+++ scripts/_helpers.py
@@ -1089,6 +1089,10 @@ def load_cutout(
     return cutout
 
 
+def make_index(c, carrier):
+    return carrier + " " + c.bus0 + " -> " + c.bus1
+
+
 def extract_grid_data_tyndp(
     links, carrier="Transmission line", replace_dict: dict = {}
 ):
@@ -1131,9 +1135,6 @@ def extract_grid_data_tyndp(
 
     # Combine into unidirectional links and return
     h2_grid = pd.concat([forward_links, reverse_links])
-
-    def make_index(c, carrier):
-        return carrier + " " + c.bus0 + " -> " + c.bus1
 
     h2_grid.index = h2_grid.apply(make_index, axis=1, args=(carrier,))